Modificacion BD y uso del cliente http

+++ database/migrations/2017_03_05_042521_add_cuentas.php
@@ -17,10 +17,10 @@ class AddCuentas extends Migration
             $table->increments('id');
             $table->string('numero');
             $table->string('tipo');
-            $table->integer('saldo_cuenta');
-            $table->integer('limite');
-            $table->integer('saldo');
-            $table->integer('cupo_disponible');
+            $table->float('saldo_cuenta');
+            $table->float('limite');
+            $table->float('saldo');
+            $table->float('cupo_disponible');
             $table->string('fecha_corte');
  //           $table->integer('id_cuenta_usuario')->unsigned();
             $table->timestamps();
